+++ scripts/entrypoint.sh
@@ -12,6 +12,7 @@ if [ $# -eq 0 ] ; then
   sh ./scripts/start_server.sh
 elif [ "${1}" = "test" ] ; then
   echo "Run Tests"
+  gunzip test/data/preindexed.sql.gz
   make test
 elif [ "${1}" = "async" ] ; then
   sh ./scripts/run_async.sh